requirements.txt
Django==3.2.4
djangorestframework==3.12.4
python-decouple==3.4
django-cors-headers==3.7.0
gunicorn==20.1.0
whitenoise==5.2.0
django-heroku==0.3.1